+++ test/runtests.jl
@@ -107,43 +107,76 @@ pnts, distM, clusts, probs, oracle_coclustering = data
 
 
     # check the sampler 
-    params = fitprior(pnts, "k-means", false; useR = false).params
-    options1 = MCMCOptionsList(usesalso = true)
-    options2 = MCMCOptionsList(usesalso = false)
-    options3 = MCMCOptionsList(pointestimation = false, usesalso = false)
     data = MCMCData(D = distM)
     @testset "Test sampler" begin
     @test (
-        try
-            result1 = runsampler(data, options1, params) # point estimation with salso
-            true
-        catch e
-            false
-        end
-    )
-    @test (
-        try
-            result2 = runsampler(data, options2, params) # point estimation without salso 
-            true
-        catch e
-            false
-        end
-    )
-    @test (
-        try
-            result3 = runsampler(data, options3, params) # no point estimation
-            sum(result3.pntestimate .== 0) == length(result3.pntestimate)
-        catch e
-            false
-        end
-    )
-    @test (
         try 
-            result4 = runsampler(data) # test defaults
+            result = runsampler(data) # test defaults
             true
         catch e
             false
         end
     )
     end
+
+    # Test point estimation
+    result = runsampler(data)
+    K = 50
+    n = 500
+    temp = sample(1:K, n)
+    @testset "Test point estimation" begin
+    @test infodist(temp, temp; normalised = true) ≈ 0 atol = 1e-9
+    @test infodist(temp, temp; normalised = false) ≈ 0 atol = 1e-9
+    @test binderloss(temp, temp; normalised = true) ≈ 0 atol = 1e-9
+    @test binderloss(temp, temp; normalised = false) ≈ 0 atol = 1e-9
+    @test (
+        try 
+            pntestimate1 = getpointestimate(result; method = "MAP")
+            true
+        catch e
+            false
+        end
+    )
+    @test (
+        try 
+            pntestimate2 = getpointestimate(result; method = "MLE")
+            true
+        catch e
+            false
+        end
+    )
+    @test (
+        try 
+            pntestimate3 = getpointestimate(result; loss = "binder", method = "MPEL")
+            true
+        catch e
+            false
+        end
+    )
+    @test (
+        try 
+            pntestimate4 = getpointestimate(result; loss = "omARI", method = "MPEL")
+            true
+        catch e
+            false
+        end
+    )
+    @test (
+        try 
+            pntestimate5 = getpointestimate(result; loss = "VI", method = "MPEL")
+            true
+        catch e
+            false
+        end
+    )
+    @test (
+        try 
+            pntestimate6 = getpointestimate(result; loss = "ID", method = "MPEL")
+            true
+        catch e
+            false
+        end
+    )
+    end
+    
 end